requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-groq
langchain-google-genai
openai
langchain-community
langchain_openai